move IpcRendererNavigationEvents to common
