+++ src/common/front-end-routing/app-navigation-channel.ts
@@ -2,7 +2,7 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import { IpcRendererNavigationEvents } from "../../renderer/navigation/events";
+import { IpcRendererNavigationEvents } from "../ipc/navigation-events";
 import type { MessageChannel } from "../utils/channel/message-channel-listener-injection-token";
 
 export type AppNavigationChannel = MessageChannel<string>;
