+++ src/common/ipc/navigation-events.ts
